+++ libraries/rotaryDecoderSwitch/rotaryDecoderSwitch.h
@@ -2,7 +2,7 @@
 //
 //    FILE: rotaryDecoderSwitch.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.2.1
+// VERSION: 0.3.0
 //    DATE: 2021-05-17
 // PURPOSE: Arduino library for rotary decoder (with switch)
 //     URL: https://github.com/RobTillaart/rotaryDecoderSwitch
@@ -11,7 +11,7 @@
 #include "Arduino.h"
 #include "Wire.h"
 
-#define ROTARY_DECODER_SWITCH_LIB_VERSION         (F("0.2.1"))
+#define ROTARY_DECODER_SWITCH_LIB_VERSION         (F("0.3.0"))
 
 
 class rotaryDecoderSwitch
@@ -33,22 +33,29 @@ public:
   bool    update();         //  assumes two directions   => +1 and -1
   bool    updateSingle();   //  assumes single direction => + ++ +++
 
+  //  re = rotary encoder
   int32_t getValue(uint8_t re);
-  void    setValue(uint8_t re, int32_t val = 0);
+  void    setValue(uint8_t re, int32_t value = 0);
   bool    isKeyPressed(uint8_t re);
+
+
+  //  READ - WRITE interface
+  uint8_t read1(uint8_t pin);
+  bool    write1(uint8_t pin, uint8_t value);
+  uint8_t read8();
+  bool    write8(uint8_t value);
+
 
   //  DEBUG
   uint8_t getLastPosition(uint8_t re);
-  uint8_t getRaw();
 
 
-private:
+protected:
   uint8_t   _count = 0;
   uint8_t   _lastValue = 0;
   uint8_t   _lastPos[2] = { 0, 0 };
   int32_t   _encoder[2] = { 0, 0 };
 
-  uint8_t   _read8();
   uint8_t   _address;
   TwoWire * _wire;
 };