+++ libraries/rotaryDecoderSwitch/CHANGELOG.md
@@ -6,8 +6,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.3.0] - 2024-01-06
+- add **read1(pin)** see #10
+- add **write1(pin, value)** experimental see #10
+- made **read8()** and **write8()** public for faster multi pin access
+- removed **getRaw()** use **read8()**
+- update readme.md
+
+----
+
 ## [0.2.1] - 2024-01-06
-- Fix URL examples
+- Fix URL in examples
 - minor edits
 
 
@@ -41,3 +50,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.1.0] - 2021-05-17
 - initial version
+
+